moved src folder back to client folder
