+++ client/src/Cartograph/__tests__/LocalShops.test.js
