+++ client/src/Cartograph/__tests__/Map.test.js
