+++ client/src/Nav/index.css
@@ -1,27 +1,20 @@
-:local(.navbar) {
-  color: #4d79ff !important;
-  /* width: 100%;
-  height: 100px;
-  font-family: Times New Roman;
-  font-size: 18px;
-  display: inline-flex;
-  justify-content: space-between; */
+.navbar {
+  background:#000066 !important;
 }
-/*
+
+.navbar-brand {
+  display: flex;
+  align-items: center;
+  position: fixed;
+  z-index: 2;
+}
+
 .nav {
-  background-color: #4d79ff;
-
+  float: right;
+  position: relative;
+  margin-left: 200px;
+  z-index: 2;
 }
-.header-nav ul {
-  list-style-type: none;
-  margin-top: 30px;
-  font-style: italic;
-}
-
-.header-nav li {
-  float: left;
-  margin: 10px;
-} */
 
 h1 {
   text-align: center;
@@ -29,11 +22,14 @@ h1 {
 }
 
 .logo {
-  width: 20%;
+  width: 8%;
   float: left;
+  position: relative;
   border: 1px solid #ddd;
   border-radius: 4px;
   padding: 0px;
+  margin-top: 0;
+  margin-right: 50px;
   background-size: cover;
   display: block;
   border-radius: 100px;
@@ -42,12 +38,12 @@ h1 {
 }
 
 .logbody {
-  position:fixed;
-  background-image: url("../Nav/cartography-2.jpg");
-  background-size: cover;
-  top: 0;
-  bottom: 0;
-  left: 0;
-  right: 0;
-  opacity: 0.2;
+	position:fixed;
+	background-image: url("../Nav/cartography-2.jpg");
+	background-size: cover;
+	top: 0;
+	bottom: 0;
+	left: 0;
+	right: 0;
+	opacity: 0.4;
 }
