+++ client/src/StockShop/__tests__/StockShop.test.js
