+++ client/src/StoreProfile/__tests__/StoreProfile.test.js
