--- src/Nav/index.css
@@ -1,49 +1,0 @@
-.navbar {
-  background:#000066 !important;
-}
-
-.navbar-brand {
-  display: flex;
-  align-items: center;
-  position: fixed;
-  z-index: 2;
-}
-
-.nav {
-  float: right;
-  position: relative;
-  margin-left: 200px;
-  z-index: 2;
-}
-
-h1 {
-  text-align: center;
-  margin-left: 100px;
-}
-
-.logo {
-  width: 8%;
-  float: left;
-  position: relative;
-  border: 1px solid #ddd;
-  border-radius: 4px;
-  padding: 0px;
-  margin-top: 0;
-  margin-right: 50px;
-  background-size: cover;
-  display: block;
-  border-radius: 100px;
-  -webkit-border-radius: 100px;
-  -moz-border-radius: 100px;
-}
-
-.logbody {
-	position:fixed;
-	background-image: url("../Nav/cartography-2.jpg");
-	background-size: cover;
-	top: 0;
-	bottom: 0;
-	left: 0;
-	right: 0;
-	opacity: 0.4;
-}
